uint256 -> uint

+++ src/contracts/Exchange.sol
@@ -6,16 +6,16 @@ import './Token.sol';
 
 contract Exchange {
 	address public feeAccount;
-	uint256 public feePercent;
+	uint public feePercent;
 	// store ether on zero-address like ether token
 	address constant ETH_ADDRESS = address(0);
 	// token address => user address => actual number of tokens
-	mapping(address => mapping(address => uint256)) public tokens; 
+	mapping(address => mapping(address => uint)) public tokens; 
 	
-	event Deposit(address token, address user, uint256 amount, uint256 balance);
-	event Withdraw(address token, address user, uint256 amount, uint256 balance);
+	event Deposit(address token, address user, uint amount, uint balance);
+	event Withdraw(address token, address user, uint amount, uint balance);
 	
-	constructor(address _feeAccount, uint256 _feePercent) public {
+	constructor(address _feeAccount, uint _feePercent) public {
 		feeAccount = _feeAccount;
 		feePercent = _feePercent;
 	} 
@@ -29,14 +29,14 @@ contract Exchange {
 		emit Deposit(ETH_ADDRESS, msg.sender, msg.value, tokens[ETH_ADDRESS][msg.sender]);
 	}
 	
-	function withdrawEther(uint256 _amount)  public {
+	function withdrawEther(uint _amount)  public {
 		require(tokens[ETH_ADDRESS][msg.sender] >= _amount);
 		tokens[ETH_ADDRESS][msg.sender] = tokens[ETH_ADDRESS][msg.sender] - _amount;
 		payable(msg.sender).transfer(_amount);
 		emit Withdraw(ETH_ADDRESS, msg.sender, _amount, tokens[ETH_ADDRESS][msg.sender]);
 	}
 	
-	function depositToken(address _token, uint256 _amount) public {
+	function depositToken(address _token, uint _amount) public {
 		require(_token != ETH_ADDRESS);
 		// Token(_token) - instance of our token on the Etherium network
 		// this - this Exchange SC
@@ -45,7 +45,7 @@ contract Exchange {
 		emit Deposit(_token, msg.sender, _amount, tokens[_token][msg.sender]);
 	}
 	
-	function withdrawToken(address _token, uint256 _amount) public {
+	function withdrawToken(address _token, uint _amount) public {
 		require(_token != ETH_ADDRESS);
     require(tokens[_token][msg.sender] >= _amount);
     tokens[_token][msg.sender] = tokens[_token][msg.sender] - _amount;
@@ -53,7 +53,7 @@ contract Exchange {
     emit Withdraw(_token, msg.sender, _amount, tokens[_token][msg.sender]);
 	}
 	
-	function balanceOf(address _token, address _user) public view returns(uint256) {
+	function balanceOf(address _token, address _user) public view returns(uint) {
 		return tokens[_token][_user];
 	}
 }
